Service _sessions Type : set<session> -> map<int,session>
PacketHandler
Handle_CS_Send_Pos()

+++ ServerLib/User.h
@@ -2,15 +2,15 @@
 #include "ServerPch.h"
 
 struct Position {
-	int32 x = 0;
-	int32 y = 0;
-	int32 z = 0;
+	float x = 0;
+	float y = 0;
+	float z = 0;
 };
 
 struct Velocity {
-	int32 x = 0;
-	int32 y = 0;
-	int32 z = 0;
+	float x = 0;
+	float y = 0;
+	float z = 0;
 };
 
 class UserInfo {
@@ -20,8 +20,8 @@ public:
 	Velocity& GetVelocity() { return _velocity; }
 
 	void SetId(int32 id) { _id = id; }
-	void SetPosition(int32 x, int32 y, int32 z) { _position.x = x; _position.y = y; _position.z = z; }
-	void SetVelocity(int32 x, int32 y, int32 z) { _velocity.x = x; _velocity.y = y; _velocity.z = z; }
+	void SetPosition(float x, float y, float z) { _position.x = x; _position.y = y; _position.z = z; }
+	void SetVelocity(float x, float y, float z) { _velocity.x = x; _velocity.y = y; _velocity.z = z; }
 
 private:
 	int32 _id;
